fix: new newsletter section
