+++ apps/comps/components/leaderboard/index.tsx
@@ -27,11 +27,15 @@ const categories = [
 export function LeaderboardSection() {
   const [selected, setSelected] = React.useState(LeaderboardTypes.TRADING);
   const [limit, setLimit] = React.useState(10);
-  const { data: leaderboard, isLoading } = useLeaderboards({
-    limit,
-    offset: 0,
-  });
-  console.log("RAFAEL", leaderboard);
+  const { data: leaderboard, isLoading } = useLeaderboards({});
+  const toRender = React.useMemo(
+    () =>
+      leaderboard && leaderboard.agents
+        ? leaderboard?.agents.slice(0, limit)
+        : [],
+    [leaderboard?.agents, limit],
+  );
+  console.log("RAFAEL", leaderboard, limit);
 
   return (
     <div className="mb-10">
@@ -128,7 +132,7 @@ export function LeaderboardSection() {
         ) : (
           <LeaderboardTable
             onExtend={() => setLimit((prev) => prev + 10)}
-            agents={leaderboard?.agents || []}
+            agents={toRender}
             loaded
           />
         )}
